Extend the plugin to get the template data from a defined object saved on the file object.
I changed the data parameter to except either an object or a string. When a string is given the plugin creates a data object and sets the object to the property with the same name attached to the file object. The data can be set bey another plugin.

I want to use a gulp workflow to process markdown files with a frontmatter for the page content and place the content into a saved html wrapper with header and footer. The frontmatter data is extracted from the markdown file and saved in a property (with a default or given name) on the file object by the frontmatter plugin. The wrap plugin can now read the frontmatter data when I set the data as a string with the property name.

Signed-off-by: uhunkler <urs.hunkler@unodo.de>