If data is an object it is a "static" value used for all files. If data is a "string" it is a link to the data which must be read from each file.
The data value is read once when the plugin is initialized.
If it is an object the same data is used for all files.
If data is a string save the datatype
and the string value. The mapped function must read the
data saved on each file in each round.

Signed-off-by: uhunkler <urs.hunkler@unodo.de>